+++ packages/graphql/tests/schema/array-methods.test.ts
@@ -98,6 +98,10 @@ describe("Arrays Methods", () => {
             input ActorActedInConnectFieldInput {
               connect: [MovieConnectInput!]
               edge: ActedInCreateInput
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: MovieConnectWhere
             }
 
@@ -362,6 +366,10 @@ describe("Arrays Methods", () => {
             input MovieActorsConnectFieldInput {
               connect: [ActorConnectInput!]
               edge: ActedInCreateInput
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: ActorConnectWhere
             }