+++ packages/graphql/tests/schema/interfaces.test.ts
@@ -156,6 +156,10 @@ describe("Interfaces", () => {
 
             input MovieNodeMoviesConnectFieldInput {
               connect: [MovieConnectInput!]
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: MovieConnectWhere
             }