+++ packages/graphql/tests/schema/issues/162.test.ts
@@ -216,6 +216,10 @@ describe("162", () => {
 
             input TigerJawLevel2Part1ConnectFieldInput {
               connect: TigerJawLevel2Part1ConnectInput
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: TigerJawLevel2Part1ConnectWhere
             }
 
@@ -324,6 +328,10 @@ describe("162", () => {
             }
 
             input TigerJawLevel2Part1TigerConnectFieldInput {
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: TigerConnectWhere
             }