+++ packages/graphql/tests/schema/query-direction.test.ts
@@ -153,6 +153,10 @@ describe("Query Direction", () => {
 
             input UserFriendsConnectFieldInput {
               connect: [UserConnectInput!]
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: UserConnectWhere
             }
 
@@ -438,6 +442,10 @@ describe("Query Direction", () => {
 
             input UserFriendsConnectFieldInput {
               connect: [UserConnectInput!]
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: UserConnectWhere
             }
 
@@ -723,6 +731,10 @@ describe("Query Direction", () => {
 
             input UserFriendsConnectFieldInput {
               connect: [UserConnectInput!]
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: UserConnectWhere
             }