+++ packages/graphql/tests/schema/string-comparators.test.ts
@@ -563,6 +563,10 @@ describe("String Comparators", () => {
             input ActorActedInConnectFieldInput {
               connect: [MovieConnectInput!]
               edge: ActedInCreateInput
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: MovieConnectWhere
             }
 
@@ -846,6 +850,10 @@ describe("String Comparators", () => {
             input MovieActorsConnectFieldInput {
               connect: [ActorConnectInput!]
               edge: ActedInCreateInput
+              \\"\\"\\"
+              Whether or not to overwrite any matching relationship with the new properties.
+              \\"\\"\\"
+              overwrite: Boolean
               where: ActorConnectWhere
             }