+++ packages/graphql/tests/tck/aggregations/where/node/duration.test.ts
@@ -68,7 +68,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN any(var2 IN collect(this1.someDuration) WHERE var2 = $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -85,8 +85,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -113,7 +112,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN any(var2 IN collect(this1.someDurationAlias) WHERE var2 = $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -130,8 +129,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -158,7 +156,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN any(var2 IN collect(this1.someDuration) WHERE var2 > $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -175,8 +173,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -203,7 +200,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN any(var2 IN collect(this1.someDuration) WHERE var2 >= $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -220,8 +217,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -248,7 +244,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN any(var2 IN collect(this1.someDuration) WHERE var2 < $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -265,8 +261,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -293,7 +288,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN any(var2 IN collect(this1.someDuration) WHERE var2 <= $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
@@ -310,8 +305,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -338,7 +332,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN avg(this1.someDuration) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -355,8 +349,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -383,7 +376,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN avg(this1.someDuration) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -400,8 +393,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -428,7 +420,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN avg(this1.someDuration) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -445,8 +437,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -473,7 +464,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN avg(this1.someDuration) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -490,8 +481,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -518,7 +508,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN avg(this1.someDuration) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -535,8 +525,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -563,7 +552,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN min(this1.someDuration) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -580,8 +569,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -608,7 +596,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN min(this1.someDuration) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -625,8 +613,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -653,7 +640,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN min(this1.someDuration) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -670,8 +657,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -698,7 +684,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN min(this1.someDuration) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -715,8 +701,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -743,7 +728,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN min(this1.someDuration) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -760,8 +745,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -788,7 +772,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN max(this1.someDuration) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -805,8 +789,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -833,7 +816,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN max(this1.someDuration) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -850,8 +833,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -878,7 +860,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN max(this1.someDuration) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -895,8 +877,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -923,7 +904,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN max(this1.someDuration) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -940,8 +921,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });
@@ -968,7 +948,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                 RETURN max(this1.someDuration) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
@@ -985,8 +965,7 @@ describe("Cypher Aggregations where node with Duration", () => {
                         \\"low\\": 0,
                         \\"high\\": 0
                     }
-                },
-                \\"param1\\": true
+                }
             }"
         `);
     });