+++ packages/graphql/tests/tck/aggregations/where/node/float.test.ts
@@ -68,14 +68,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN any(var2 IN collect(this1.someFloat) WHERE var2 = $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -102,14 +101,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN any(var2 IN collect(this1.someFloatAlias) WHERE var2 = $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -136,14 +134,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN any(var2 IN collect(this1.someFloat) WHERE var2 > $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -170,14 +167,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN any(var2 IN collect(this1.someFloat) WHERE var2 >= $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -204,14 +200,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN any(var2 IN collect(this1.someFloat) WHERE var2 < $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -238,14 +233,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN any(var2 IN collect(this1.someFloat) WHERE var2 <= $param0) AS var3
             }
             WITH *
-            WHERE var3 = $param1
+            WHERE var3 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -272,14 +266,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN avg(this1.someFloat) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -306,14 +299,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN avg(this1.someFloat) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -340,14 +332,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN avg(this1.someFloat) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -374,14 +365,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN avg(this1.someFloat) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -408,14 +398,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN avg(this1.someFloat) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -442,14 +431,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN sum(this1.someFloat) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -476,14 +464,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN sum(this1.someFloat) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -510,14 +497,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN sum(this1.someFloat) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -544,14 +530,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN sum(this1.someFloat) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -578,14 +563,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN sum(this1.someFloat) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -612,14 +596,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN min(this1.someFloat) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -646,14 +629,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN min(this1.someFloat) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -680,14 +662,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN min(this1.someFloat) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -714,14 +695,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN min(this1.someFloat) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -748,14 +728,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN min(this1.someFloat) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -782,14 +761,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN max(this1.someFloat) = $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -816,14 +794,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN max(this1.someFloat) > $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -850,14 +827,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN max(this1.someFloat) >= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -884,14 +860,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN max(this1.someFloat) < $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });
@@ -918,14 +893,13 @@ describe("Cypher Aggregations where node with Float", () => {
                 RETURN max(this1.someFloat) <= $param0 AS var2
             }
             WITH *
-            WHERE var2 = $param1
+            WHERE var2 = true
             RETURN this { .content } AS this"
         `);
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`
             "{
-                \\"param0\\": 10,
-                \\"param1\\": true
+                \\"param0\\": 10
             }"
         `);
     });